Added a version of project_aoi function that can take gtapv7_r251_label as column name when filtering AOI
This version of the function make disputed lands to project AOI. I also added a local function that is similar to hb.extract_features_in_shapefile_by_attribute but it can dissolve features if the length of features is greater than 1. Later we may consider adding this fucntion directly yo hb.